expose avg and count
