+++ velox/functions/prestosql/aggregates/CMakeLists.txt
@@ -24,6 +24,7 @@ add_library(
   ArbitraryAggregate.cpp
   ArrayAggAggregate.cpp
   AverageAggregate.cpp
+  AverageAggregate.h
   BitwiseAggregates.cpp
   BoolAggregates.cpp
   CountIfAggregate.cpp
@@ -33,6 +34,7 @@ add_library(
   MinMaxAggregates.cpp
   MinMaxByAggregates.cpp
   CountAggregate.cpp
+  CountAggregate.h
   PrestoHasher.cpp
   SingleValueAccumulator.cpp
   SumAggregate.cpp
